// src/types/db.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
// ...
      courses: {
        Row: {
          created_at: string | null
          id: string
          level: string | null
          name: string
          position: number | null
        }
        Insert: {
          created_at?: string | null
          id?: string
          level?: string | null
          name: string
          position?: number | null
        }
        Update: {
          created_at?: string | null
          id?: string
          level?: string | null
          name?: string
          position?: number | null
        }
        Relationships: []
      }
// ...
      students: {
        Row: {
          course_id: string | null
          created_at: string | null
          full_name: string
          id: string
          rut: string | null
        }
        Insert: {
          course_id?: string | null
          created_at?: string | null
          full_name: string
          id?: string
          rut?: string | null
        }
        Update: {
          course_id?: string | null
          created_at?: string | null
          full_name?: string
          id?: string
          rut?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "students_course_id_fkey"
            columns: ["course_id"]
            isOneToOne: false
            referencedRelation: "courses"
            referencedColumns: ["id"]
          },
        ]
      }
      tests: {
        Row: {
          course_id: string | null
          created_at: string | null
          date: string
          description: string | null
          id: string
          subject: string
          type: string
        }
        Insert: {
          course_id?: string | null
          created_at?: string | null
          date: string
          description?: string | null
          id?: string
          subject: string
          type: string
        }
        Update: {
          course_id?: string | null
          created_at?: string | null
          date?: string
          description?: string | null
          id?: string
          subject?: string
          type?: string
        }
        Relationships: [
          {
            foreignKeyName: "tests_course_id_fkey"
            columns: ["course_id"]
            isOneToOne: false
            referencedRelation: "courses"
            referencedColumns: ["id"]
          },
        ]
      }
    }
// ...
  }
}